typefix to HSLAPixel

+++ lab_debug/cs225/HSLAPixel.cpp
@@ -19,14 +19,14 @@ namespace cs225 {
     a = 1.0;
   }
 
-  HSLAPixel::HSLAPixel(float hue, float saturation, float luminance) {
+  HSLAPixel::HSLAPixel(double hue, double saturation, double luminance) {
     h = hue;
     s = saturation;
     l = luminance;
     a = 1.0;
   }
 
-  HSLAPixel::HSLAPixel(float hue, float saturation, float luminance, float alpha) {
+  HSLAPixel::HSLAPixel(double hue, double saturation, double luminance, double alpha) {
     h = hue;
     s = saturation;
     l = luminance;
